Convert to type-generic math (#2050)
* [Optimization] Convert to type-generic math

* add std:: prefix in obj-c++ files

* more cleanup

* revert test changes

* convert min and max back to fmin/fmax

+++ AsyncDisplayKit/ASEditableTextNode.mm
@@ -11,6 +11,7 @@
 #import "ASEditableTextNode.h"
 
 #import <objc/message.h>
+#import <tgmath.h>
 
 #import "ASDisplayNode+Subclasses.h"
 #import "ASEqualityHelpers.h"
@@ -238,9 +239,9 @@
 {
   ASTextKitComponents *displayedComponents = [self isDisplayingPlaceholder] ? _placeholderTextKitComponents : _textKitComponents;
   CGSize textSize = [displayedComponents sizeForConstrainedWidth:constrainedSize.width];
-  CGFloat width = ceilf(textSize.width + _textContainerInset.left + _textContainerInset.right);
-  CGFloat height = ceilf(textSize.height + _textContainerInset.top + _textContainerInset.bottom);
-  return CGSizeMake(fminf(width, constrainedSize.width), fminf(height, constrainedSize.height));
+  CGFloat width = std::ceil(textSize.width + _textContainerInset.left + _textContainerInset.right);
+  CGFloat height = std::ceil(textSize.height + _textContainerInset.top + _textContainerInset.bottom);
+  return CGSizeMake(std::fmin(width, constrainedSize.width), std::fmin(height, constrainedSize.height));
 }
 
 - (void)layout
